requirements.txt
fastapi
uvicorn
pymilvus
requests
python-dotenv
pypdf
python-docx
python-pptx
openpyxl
pandas
tiktoken
